Done basic product controller functions

+++ src/main/java/pl/coderslab/entity/Meal.java
@@ -23,5 +23,5 @@ public class Meal {
     private double totalProtein;
     private double totalCarbohydrates;
     private double totalFat;
-    private double totalCalories;
+    private int totalCalories;
 }
